pyproject.toml
[project]
name = "lissen-user-journey"
version = "0.1.0"
description = "User Journey Analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-cloud-bigquery",
  "pandas",
  "pyyaml",
  "streamlit",
  "plotly",
  "db-dtypes"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "pre-commit"
]

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = [".venv/*", "__pycache__/*"]
